Build fix for Linux x86
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@6609 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Core/VideoCommon/Src/OpenCL/OCLTextureDecoder.cpp
@@ -93,13 +93,13 @@ void TexDecoder_OpenCL_Initialize()
 			return;
 
 		cl_int err = 1;
-		u64 binary_size = 0;
+		size_t binary_size = 0;
 		char *binary = NULL;
 		char *header = NULL;
-		u32 nDevices = 0;
+		size_t nDevices = 0;
 		cl_device_id *devices = NULL;
 		cl_program program = NULL;
-		u64 *binary_sizes = NULL;
+		size_t *binary_sizes = NULL;
 		char **binaries = NULL;		
 		char filename[1024];
